backend/requirements.txt
# Web框架
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Google AI SDK
google-generativeai==0.8.3

# LangChain 核心
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-community==0.3.13
langchain-text-splitters>=0.3.3

# 向量数据库
langchain-chroma==0.1.4
chromadb==0.5.23

# 文档处理
pypdf==5.1.0              # PDF处理
python-docx==1.1.2        # Word处理

# 环境变量
python-dotenv==1.0.1

# 工具
pydantic==2.10.5

# 本地 Embedding 模型（新增）
sentence-transformers==3.3.1
langchain-huggingface==0.1.2